requirements.txt
asgiref==3.7.2
Django==5.0.3
django-ipware==6.0.4
python-ipware==2.0.2
sqlparse==0.4.4
tzdata==2024.1
